backend/requirements.txt
flask~=3.0.2
Flask-SocketIO~=5.3.6
boto3~=1.34.103
PyJWT~=2.8.0
cryptography~=42.0.4
requests~=2.31.0 